GRIDEDIT-213: for consistency with triangle interpolation, rename Averaging to AveragingInterpolation
