+++ src/AveragingInterpolation.hpp
